Mark app_dummy deprecated.
This isn't needed any more. We'll mark it deprecated to encourage
people to drop it in favor of letting `LOCAL_WHOLE_STATIC_LIBRARIES`
do its job, but the sample didn't tell people to do this until
recently, so doing anything more than that would require a lot of
churn in user makefiles.

Test: ./run_tests.py --filter native_app_glue
Bug: https://github.com/android-ndk/ndk/issues/381
Change-Id: I2be0f0676d90846bfa171670e512dbe7ebf4c68c